Parse dates to turn them into sortable ISO dates
I may need the uncertainly data back in the application at some point,
but for now I need definate numbers for every date so that I can
properly sort the images on a timeline. It doesn't make sense to do this
in the client everytime on page load so we normalize the data here.

Eventually I may need to move this into another column so that I can
keep the uncertainy date.