Highlight the clicked image on page render
When you come from the locations page, you should see the image that you
clicked highlighted. This passes an argument in the URL and then sets
the image display based on the DOM x position of the image. It's a bit
of a hack for now, but will be improved later.