Cleaning up highlight function
This breaks the key aspects of the highlight operation into their own
functions and fixes a few style issues. This is important to do because
we'll be adding another highlight function that is based on index of the
photo, not the cursor position. This will help with the next and
previosu buttons.